+++ Rapid/include/rapid/IO/systemExecute.h
@@ -22,7 +22,7 @@ namespace rapid
 			{
 				throw std::runtime_error("popen() failed!");
 			}
-			while (fgets(buffer.data(), buffer.size(), pipe.get()) != nullptr)
+			while (fgets(buffer.data(), (int) buffer.size(), pipe.get()) != nullptr)
 			{
 				result += buffer.data();
 			}
